abstract code to make it shared between Audio and video components since they work the same
